+++ esp-radio/src/radio_clocks/clocks_ll/esp32h2.rs
@@ -1,7 +1,12 @@
 fn ble_ieee802154_clock_enable(en: bool) {
     regs!(MODEM_SYSCON).clk_conf().modify(|_, w| {
         w.clk_zb_apb_en().bit(en);
-        w.clk_zb_mac_en().bit(en)
+        w.clk_zb_mac_en().bit(en);
+        w.clk_modem_sec_en().bit(en);
+        w.clk_modem_sec_ecb_en().bit(en);
+        w.clk_modem_sec_ccm_en().bit(en);
+        w.clk_modem_sec_bah_en().bit(en);
+        w.clk_modem_sec_apb_en().bit(en)
     });
 
     regs!(MODEM_SYSCON).clk_conf1().modify(|_, w| {